hopefully last mypy-black problem
